backend/pyproject.toml
[project]
name = "public-sector-backend"
version = "0.1.0"
description = "AI-Powered Public Sector Document Processing Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "celery>=5.4.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14.0",
    "psycopg[binary]>=3.2.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
    "dashscope>=1.22.0",
    "oss2>=2.19.0",
    "python-jose[cryptography]>=3.3.0",
    "redis>=5.2.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "httpx>=0.28.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "COM", "C4", "T20", "RET", "SIM"]
ignore = ["S101", "COM812", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
